Clean up tests
Also fixed a bug where all cards were considered to be unknown 1s.
