+++ test/h-group/level-4.js
@@ -5,11 +5,12 @@ import { describe, it } from 'node:test';
 
 import { PLAYER, setup } from '../test-utils.js';
 import HGroup from '../../src/conventions/h-group.js';
-import { ACTION } from '../../src/constants.js';
+import { ACTION, CLUE } from '../../src/constants.js';
 import * as Utils from '../../src/util.js';
 import logger from '../../src/logger.js';
 
 import { find_clues } from '../../src/conventions/h-group/clue-finder/clue-finder.js';
+import { determine_playable_card, find_urgent_actions } from '../../src/conventions/h-group/action-helper.js';
 
 logger.setLevel(logger.LEVELS.ERROR);
 
@@ -54,5 +55,62 @@ describe('trash chop move', () => {
 
 		const { save_clues } = find_clues(state);
 		assert(save_clues[PLAYER.BOB] === undefined);
+	});
+});
+
+describe('order chop move', () => {
+	it('will find an ocm to the next player', () => {
+		const state = setup(HGroup, [
+			['xx', 'xx', 'xx', 'xx', 'xx'],
+			['r4', 'r4', 'g4', 'r3', 'r5']
+		], 4);
+
+		// Bob clues Alice 1, touching slots 3 and 4.
+		const action = { type: 'clue', clue: { type: CLUE.RANK, value: 1 }, giver: PLAYER.ALICE, list: [1, 2], target: PLAYER.ALICE, turn: 0 };
+		state.handle_action(action);
+
+		const our_hand = state.hands[state.ourPlayerIndex];
+
+		const playable_priorities = determine_playable_card(state, [our_hand[2], our_hand[3]]);
+		const { play_clues, save_clues, fix_clues } = find_clues(state);
+		const urgent_actions = find_urgent_actions(state, play_clues, save_clues, fix_clues, playable_priorities);
+
+		assert.equal(urgent_actions[1][0].type, ACTION.PLAY);
+		assert.equal(urgent_actions[1][0].target, our_hand[2].order);
+	});
+
+	it('will find an ocm to cathy', () => {
+		const state = setup(HGroup, [
+			['xx', 'xx', 'xx', 'xx', 'xx'],
+			['g3', 'r3', 'y3', 'y4', 'y4'],
+			['r4', 'r4', 'g4', 'r3', 'r5'],
+		], 4);
+
+		// Bob clues Alice 1, touching slots 2, 3 and 4.
+		const action = { type: 'clue', clue: { type: CLUE.RANK, value: 1 }, giver: PLAYER.ALICE, list: [1, 2, 3], target: PLAYER.ALICE, turn: 0 };
+		state.handle_action(action);
+
+		const our_hand = state.hands[PLAYER.ALICE];
+
+		const playable_priorities = determine_playable_card(state, [our_hand[1], our_hand[2], our_hand[3]]);
+		const { play_clues, save_clues, fix_clues } = find_clues(state);
+		const urgent_actions = find_urgent_actions(state, play_clues, save_clues, fix_clues, playable_priorities);
+
+		assert.equal(urgent_actions[5][0].type, ACTION.PLAY);
+		assert.equal(urgent_actions[5][0].target, our_hand[1].order);
+	});
+
+	it('will not give an ocm putting a critical on chop', () => {
+		const state = setup(HGroup, [
+			['xx', 'xx', 'xx', 'xx', 'xx'],
+			['r4', 'r4', 'g4', 'r5', 'r5']
+		], 4);
+
+		// Bob clues Alice 1, touching slots 3 and 4.
+		const action = { type: 'clue', clue: { type: CLUE.RANK, value: 1 }, giver: PLAYER.ALICE, list: [1, 2], target: PLAYER.ALICE, turn: 0 };
+		state.handle_action(action);
+
+		const { save_clues } = find_clues(state);
+		assert.deepEqual(Utils.objPick(save_clues[PLAYER.BOB], ['type', 'value']), { type: ACTION.RANK, value: 5 });
 	});
 });
